Fix several dependency cycles # Delete this line to force CI to run the JVM tests. [ci skip-jvm-tests]
