+++ src/python/pants/build_graph/BUILD
@@ -2,12 +2,13 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_library(
+  sources=["*.py", "!address.py", "!bundle.py", "!*_test.py"],
   dependencies = [
+    ':address',
+    ':bundle',
     '3rdparty/python:dataclasses',
     'src/python/pants/base:build_environment',
-    'src/python/pants/base:build_file',
     'src/python/pants/base:build_file_target_factory',
-    'src/python/pants/base:deprecated',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:fingerprint_strategy',
     'src/python/pants/base:hash_utils',
@@ -16,17 +17,47 @@ python_library(
     'src/python/pants/base:payload_field',
     'src/python/pants/base:specs',
     'src/python/pants/base:validation',
+    'src/python/pants/engine:rules',
     'src/python/pants/engine:target',
     'src/python/pants/fs',
     'src/python/pants/option',
     'src/python/pants/source',
     'src/python/pants/source:payload_fields',
     'src/python/pants/subsystem',
+    # 'src/python/pants/task',  # dep cycle
     'src/python/pants/util:dirutil',
+    'src/python/pants/util:fileutil',
+    'src/python/pants/util:frozendict',
     'src/python/pants/util:memo',
     'src/python/pants/util:meta',
     'src/python/pants/util:netrc',
+    'src/python/pants/util:objects',
     'src/python/pants/util:ordered_set',
+  ],
+  tags = {"partially_type_checked"},
+)
+
+python_library(
+  name="address",
+  sources=["address.py"],
+  dependencies = [
+    'src/python/pants/base:build_file',
+    'src/python/pants/base:deprecated',
+    'src/python/pants/util:dirutil',
+    'src/python/pants/util:strutil',
+  ],
+  tags = {"type_checked"},
+)
+
+python_library(
+  name="bundle",
+  sources=["bundle.py"],
+  dependencies=[
+    'src/python/pants/base:build_environment',
+    'src/python/pants/base:payload_field',
+    'src/python/pants/util:collections',
+    'src/python/pants/util:dirutil',
+    'src/python/pants/util:memo',
   ],
   tags = {"partially_type_checked"},
 )
